requirements.txt
git+git://github.com/heroku/kafka-helper.git@70aa539540e7d6517f0c3c1900617767ab198b6c
ipython==6.1.0
kafka-python==1.3.4
jupyter==1.0.0
